Community workflows (nccl) cant be deleted

+++ cmd/delete/workflow.sh
@@ -95,25 +95,30 @@ fi
 github_branch=$(cat $WORKFLOWS_USER_PATH/GITHUB_PUSH_BRANCH)
 
 # delete 
-if [[ -d "$WORKFLOWS_USER_PATH/$name" ]]; then
-  # delete workflow and push
-  cd "$WORKFLOWS_USER_PATH"
-  if git ls-tree -r --name-only "$github_branch" -- "$name" | grep -q .; then
-    rm -rf -- "$WORKFLOWS_USER_PATH/$name"
+target="$(readlink -f "$ODEV_PATH/cmd/new/$name.sh")"
+if [[ "$target" == "$WORKFLOWS_USER_PATH/"* ]]; then
+  if [[ -d "$WORKFLOWS_USER_PATH/$name" ]]; then
+    # delete workflow and push
+    cd "$WORKFLOWS_USER_PATH"
+    if git ls-tree -r --name-only "$github_branch" -- "$name" | grep -q .; then
+      rm -rf -- "$WORKFLOWS_USER_PATH/$name"
 
-    git add -A
-    git commit -m "Delete workflow $name"
-    git push origin "$github_branch"
-  else
-    rm -rf -- "$WORKFLOWS_USER_PATH/$name"
+      git add -A
+      git commit -m "Delete workflow $name"
+      git push origin "$github_branch"
+    else
+      rm -rf -- "$WORKFLOWS_USER_PATH/$name"
+    fi
+
+    # delete symlinks
+    sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/new/$name.sh"
+    sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/build/$name.sh"
+    sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/program/$name.sh"
+    sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/run/$name.sh"
+    sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/validate/$name.sh"
+    sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/delete/$name.sh"
   fi
-
-  # delete symlinks
-  sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/new/$name.sh"
-  sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/build/$name.sh"
-  sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/program/$name.sh"
-  sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/run/$name.sh"
-  sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/validate/$name.sh"
-  sudo "$ODEV_PATH/src/rm.sh" "$ODEV_PATH" "$ODEV_PATH/cmd/delete/$name.sh"
+  exit 1
+else
+  echo "Workflow cannot be deleted: $name"
 fi
-exit 1
